feat: export util functions and add string helper
