+++ src/utils/string.ts
@@ -1,0 +1,1 @@
+export const toCapitalize = (str: string) => str.charAt(0).toUpperCase() + str.slice(1)
